test6-databases/api/pyproject.toml
[project]
name = "test6-databases"
version = "0.1.0"
description = "RDS Postgres + DynamoDB experiments — experiment 6"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "psycopg2-binary",
    "boto3",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
